requirements.txt
paho-mqtt
pyserial
adafruit-io